Split FSAI_precond and Nys_precond for future use
