+++ examples/AFN_precond/common.make
@@ -31,7 +31,7 @@ endif
 C_SRCS 	= $(wildcard *.c)
 C_OBJS  = $(C_SRCS:.c=.c.o)
 EXES    = test_AFN.exe test_Nys.exe test_FSAI.exe
-SHARED_OBJS = ../PCG/pcg.c.o 
+SHARED_OBJS = FSAI_precond.c.o Nys_precond.c.o ../PCG/pcg.c.o 
 
 # Delete the default old-fashion double-suffix rules
 .SUFFIXES:
